requirements.txt
python>=3.7
pytorch>=0.4.0
tqdm
numpy
pyyaml
sentencepiece>=0.1.82
tensorboardX
mosestokenizer>=1.0.0
sacrebleu>=1.3.2
